requirements.txt
python-telegram-bot[job-queue]==20.7
openai
requests
python-dotenv